Refreshing 3 sec. EEPROM connected, Joystick added, NTC sensor connected

+++ Incubator/Core/Src/stm32f1xx_it.c
@@ -42,7 +42,7 @@
 
 /* Private variables ---------------------------------------------------------*/
 /* USER CODE BEGIN PV */
-volatile uint32_t g_TimestampInMilliseconds = 0;
+volatile int64_t g_TimestampInMilliseconds = 0;
 /* USER CODE END PV */
 
 /* Private function prototypes -----------------------------------------------*/
